+++ src/CMakeLists.txt
@@ -31,7 +31,7 @@
 # =============================================================================
 
 set(RASCAL_SOURCES
-    rascal/json_io.cc
+    rascal/utils/json_io.cc
     rascal/utils/units.cc
     rascal/utils/utils.cc
     rascal/utils/sparsify_fps.cc
