+++ tests/test_calculator.hh
@@ -34,7 +34,6 @@
 #include "test_structure.hh"
 
 #include "rascal/atomic_structure.hh"
-#include "rascal/json_io.hh"
 #include "rascal/representations/calculator_base.hh"
 #include "rascal/representations/calculator_sorted_coulomb.hh"
 #include "rascal/representations/calculator_spherical_covariants.hh"
@@ -42,6 +41,7 @@
 #include "rascal/representations/calculator_spherical_invariants.hh"
 #include "rascal/structure_managers/cluster_ref_key.hh"
 #include "rascal/structure_managers/structure_manager_collection.hh"
+#include "rascal/utils/json_io.hh"
 #include "rascal/utils/utils.hh"
 
 #include <memory>
